snowboard_scraper_starter/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.2
pyyaml>=6.0.1
pandas>=2.2.2
tqdm>=4.66.4
